+++ CHANGELOG.md
@@ -2346,3 +2346,4 @@
 - [2026-06-25 23:01] perf(ui): apply RepaintBoundary on animated ui widgets
 - [2026-06-25 23:14] fix(cache): prevent double tap duplicate action in cache
 - [2026-06-26 09:05] perf(i18n): optimize memory consumption in i18n list viewport
+- [2026-06-26 09:28] feat(state): add async state stream provider